move tests from test_datetime per reviewer request
